* updated data schema * read currently available params into dataobjects * reorg reduction algorithm to be more dynamic * map ingested data to algo inputs * ui button now commences reduction for real run

+++ src/snapred/backend/dao/StateId.py
@@ -11,8 +11,8 @@ class StateId:
 
     # Round inputs to reduced number of possible states
     def __init__(self, vdet_arc1: float, vdet_arc2: float, WavelengthUserReq: float, Frequency: float, Pos: int):
-        self.vdet_arc1 = round(vdet_arc1 * 2) / 2
-        self.vdet_arc2 = round(vdet_arc2 * 2) / 2
-        self.WavelengthUserReq = round(WavelengthUserReq, 1)
-        self.Frequency = round(Frequency)
-        self.Pos = Pos
+        self.vdet_arc1 = float(round(vdet_arc1 * 2) / 2)
+        self.vdet_arc2 = float(round(vdet_arc2 * 2) / 2)
+        self.WavelengthUserReq = float(round(WavelengthUserReq, 1))
+        self.Frequency = int(round(Frequency))
+        self.Pos = int(Pos)
